int to off_t for fseeks and ftells (https://trac.osgeo.org/grass/ticket/1971, merge https://trac.osgeo.org/grass/changeset/56422 from devbr6)
git-svn-id: https://svn.osgeo.org/grass/grass/branches/releasebranch_6_4@56590 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/gisdefs.h
@@ -19,7 +19,7 @@
 
 /*============================= Include Files ==============================*/
 
-/* none */
+#include <sys/types.h>	/* for off_t */
 
 /*=========================== Constants/Defines ============================*/
 
@@ -1085,8 +1085,8 @@ void G_rotate_around_point(double, double, double *, double *, double);
 void G_rotate_around_point_int(int, int, int *, int *, double);
 
 /* seek.c */
-int G_ftell(FILE *);
-void G_fseek(FILE *, int, int);
+off_t G_ftell(FILE *);
+void G_fseek(FILE *, off_t, int);
 
 /* sample.c */
 DCELL G_get_raster_sample_nearest(
